Объединил приложения content и reviews в одно и исправил свои замечания в моделях
