+++ api_yamdb/reviews/constants.py
@@ -1,0 +1,2 @@
+MAX_LENGTH_NAME = 256
+MAX_LENGTH_SLUG = 50
