+++ api_yamdb/reviews/management/commads/load_categories.py
