+++ api_yamdb/reviews/models.py
@@ -1,10 +1,88 @@
+import datetime
+
 from django.contrib.auth import get_user_model
 from django.core.validators import MaxValueValidator
 from django.db import models
 
-from content.models import Title
+from .constants import MAX_LENGTH_NAME, MAX_LENGTH_SLUG
 
 User = get_user_model()
+
+
+class NamedSlugModel(models.Model):
+    """Абстрактная модель с полями name и slug."""
+
+    name = models.CharField(
+        max_length=MAX_LENGTH_NAME,
+        unique=True,
+        verbose_name='Название'
+    )
+    slug = models.SlugField(
+        max_length=MAX_LENGTH_SLUG,
+        unique=True,
+        verbose_name='Slug'
+    )
+
+    class Meta:
+        abstract = True
+        ordering = ('name',)
+
+    def __str__(self):
+        return self.name
+
+
+class Category(NamedSlugModel):
+    """Категория."""
+
+    class Meta:
+        verbose_name = 'Категория'
+        verbose_name_plural = 'Категории'
+
+
+class Genre(NamedSlugModel):
+    """Жанр."""
+
+    class Meta:
+        verbose_name = 'Жанр'
+        verbose_name_plural = 'Жанры'
+
+
+class Title(models.Model):
+    """Произведение."""
+
+    name = models.CharField(
+        max_length=MAX_LENGTH_NAME,
+        verbose_name='Название'
+    )
+    year = models.PositiveSmallIntegerField(
+        verbose_name='Год',
+        validators=[MaxValueValidator(datetime.datetime.now().year)]
+    )
+    description = models.TextField(
+        blank=True,
+        verbose_name='Описание'
+    )
+    genre = models.ManyToManyField(
+        Genre,
+        related_name='titles',
+        verbose_name='Жанры'
+    )
+    category = models.ForeignKey(
+        Category,
+        on_delete=models.SET_NULL,
+        null=True,
+        blank=True,
+        related_name='titles',
+        verbose_name='Категория'
+    )
+
+    class Meta:
+        ordering = ('name',)
+        verbose_name = 'Произведение'
+        verbose_name_plural = 'Произведения'
+
+    def __str__(self):
+        return self.name
 
 
 class BaseModel(models.Model):
